move visibleAnswersWithGradings() from app/app/answers/get_answer.go into app/api/answers/get_current.go and split it into parts
