+++ app/api/answers/get_current.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/go-chi/render"
 
+	"github.com/France-ioi/AlgoreaBackend/app/database"
 	"github.com/France-ioi/AlgoreaBackend/app/service"
 )
 
@@ -78,4 +79,16 @@ func (srv *Service) getCurrentAnswer(rw http.ResponseWriter, httpReq *http.Reque
 
 	render.Respond(rw, httpReq, convertedResult)
 	return service.NoError
+}
+
+func visibleAnswersWithGradings(store *database.DataStore, user *database.User) *database.DB {
+	return withGradings(store.Answers().Visible(user))
+}
+
+func withGradings(answersQuery *database.DB) *database.DB {
+	return answersQuery.
+		Joins("LEFT JOIN gradings ON gradings.answer_id = answers.id").
+		Select(`answers.id, answers.author_id, answers.item_id, answers.attempt_id, answers.participant_id,
+			answers.type, answers.state, answers.answer, answers.created_at, gradings.score,
+			gradings.graded_at`)
 }
